+++ checks/domain/template-linux/input.nix
@@ -3,5 +3,5 @@ lib: lib.domain.templates.linux
   name = "test-q35";
   uuid = "2904419d-b283-4cfd-9f2c-7c3713ff809f";
   memory = { count = 6; unit = "GiB"; };
-  storage_vol_path = /Storage/MyHD.qcow2;
+  storage_vol = /Storage/MyHD.qcow2;
 }